adding delete functionality for user

+++ DigitalLifeBooks/Services/Delete.asmx.cs
@@ -5,6 +5,7 @@ using System.Web;
 using System.Web.Services;
 using DigitalLifeBooks.Models;
 using DigitalLifeBooks.AssetManagement;
+using System.Web.Security;
 
 namespace DigitalLifeBooks.Services
 {
@@ -67,7 +68,16 @@ namespace DigitalLifeBooks.Services
         private void DeleteUser(long EnityID)
         {
             var user = DataContext.Users.Single(u => u.ID == EnityID);
+            if (user.LoginName == HttpContext.Current.User.Identity.Name)
+                throw new InvalidOperationException("cannot delete yourself");
+                
+            for (int i = 0; user.Children.Count != 0; i++)
+            {
+                var child = user.Children.ElementAt(i);
+                user.Children.Remove(child);
+            }
             DataContext.Users.DeleteObject(user);
+            Membership.DeleteUser(user.LoginName);
         }
 
         private void DeleteAsset(long EnityID)
